Move GraphQL context keys out of internal package
